Fix for #2336.
Previous fixes produced a second issue: The message being sent to the error page could not be styled by the sender in any way. <b> <i> etc would not work.

This fix manages to do a case insensitive replacement of any incoming script tags, turning them into noscript tags. The result is that any browser which understands script tags simply hides them, and any browser which does not remains unaffected.

All in all, this solution looks to be the best possible solution to the problem. It does use regular expressions, but it was the only way I could find to keep the code simple.